# swagger/swagger.json
{
    "swagger": "2.0",
    "info": {
      "title": "Backend API",
      "description": "This is Backend API.",
      "version": "1.0.0"
    },
    "basePath":"/",
    "schemes": [
      "http"
    ],
    "securityDefinitions":{
      "bearerAuth":{
        "type":"apiKey",
        "name":"Authorization",
        "scheme":"Bearer",
        "in":"header",
        "description": "For accessing the API a valid JWT token must be passed in all the queries in the 'Authorization' header"
      }

    },
    "paths": {
      "/api/v1/index": {
        "get": {
          "tags": [
              "index"
          ],
          "security":[
            {
              "bearerAuth":[]
            }
          ],
          "summary": "Get list of index",
          "operationId": "get index",
          "parameters": [
          {
              "name": "limit",
              "in": "query",
              "description": "Number of results to return per page.",
              "required": false,
              "type": "integer",
              "items": {
                  "type": "integer"
              },
              "collectionFormat": "csv"
          },
          {
              "name": "offset",
              "in": "query",
              "description": "The initial index from which to return the results.",
              "required": false,
              "type": "integer",
              "items": {
                  "type": "integer"
              },
              "collectionFormat": "csv"
          }
        ],
        "responses": {
          "200": {
            "schema": {
              "$ref": "#/definitions/htmlcreate",
              "type": "string"
            }
          }
        }
        }, 
        "post": {
          "tags": [
            "index"
          ],
          "security":[
            {
              "bearerAuth":[]
            }
          ],
          "summary": "Create a index",
          "operationId": "create index",
          "parameters": [
            {
              "name": "data",
              "in": "body",
              "required": true,
              "schema": {
                "$ref": "#/definitions/htmlcreate"
              }
            }
          ],
          "responses": {
            "201": {
              "schema": {
                "$ref": "#/definitions/htmlcreate"
              }
            },
            "400": {
              "schema": {
                "$ref": "#/definitions/Error"
              }
            }
          }
        }
      }
    },
    "responses":{
      "NotFound":{
        "description":"The specified resource was not found"
      }
    },
    "definitions": {
      "htmlcreate": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "html": {
            "type": "string"
          }
        }
      },
      "pagecreate": {
        "type": "object",
        "properties": {
          "section_id": {
            "type": "string"
          },
          "sec_order": {
            "type": "integer"
          },
          "section_type": {
            "type": "string"
          },
          "section_props": {
            "type": "string"
          },
          "section_props_previous": {
            "type": "string"
          }
        }
      },
      "pageupdate": {
        "type": "object",
        "properties": {
          "sec_order": {
            "type": "integer"
          },
          "section_type": {
            "type": "string"
          },
          "is_active":{
            "type": "integer"
          },
          "scheduler_state":{
            "type": "integer"
          },
          "section_props": {
            "type": "string"
          },
          "section_props_previous": {
            "type": "string"
          }
        }
      },
      "pageupdatebulk": {
        "type": "object",
        "properties": {
          "query": {
            "type": "string"
          }
        }
      },
      "fileupload": {
        "type": "object",
        "properties": {
          "file": {
            "type": "string"
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          }
        }
      },
      "Deleted": {
        "type": "object",
        "properties": {
          "result": {
            "type": "string"
          }
        }
      },
      "ApiResponse": {
        "type": "object",
        "properties": {
          "code": {
            "type": "integer",
            "format": "int32"
          },
          "type": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "UnauthorizedError":{
        "type": "object",
        "properties": {
          "description": {
            "type": "string"
          }
        }
      }
    }
  }